refactor: Admin 기능 API는 별도의 클래스로 분리 및 url 변경
