+++ src/test/java/com/playkuround/playkuroundserver/domain/user/application/UserLoginServiceTest.java
@@ -4,6 +4,7 @@ import com.playkuround.playkuroundserver.domain.auth.token.application.TokenMana
 import com.playkuround.playkuroundserver.domain.auth.token.application.TokenService;
 import com.playkuround.playkuroundserver.domain.auth.token.dto.TokenDto;
 import com.playkuround.playkuroundserver.domain.user.dao.UserRepository;
+import com.playkuround.playkuroundserver.domain.user.exception.UserNotFoundException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -14,6 +15,7 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import java.util.Date;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.when;
 
@@ -56,6 +58,18 @@ class UserLoginServiceTest {
         assertThat(response.getGrantType()).isEqualTo(tokenDto.getGrantType());
         assertThat(response.getAccessToken()).isEqualTo(tokenDto.getAccessToken());
         assertThat(response.getRefreshToken()).isEqualTo(tokenDto.getRefreshToken());
+    }
+
+    @Test
+    @DisplayName("해당 이메일로 가입된 유저가 없다면 에러가 발생한다.")
+    void signupFail() {
+        // given
+        String email = "tester@konkuk.ac.kr";
+        when(userRepository.existsByEmail(email)).thenReturn(false);
+
+        // expect
+        assertThatThrownBy(() -> userLoginService.login(email))
+                .isInstanceOf(UserNotFoundException.class);
     }
 
 }
