+++ src/test/java/com/playkuround/playkuroundserver/domain/user/application/UserRegisterServiceTest.java
@@ -8,10 +8,13 @@ import com.playkuround.playkuroundserver.domain.user.dto.UserRegisterDto;
 import com.playkuround.playkuroundserver.domain.user.dto.response.UserRegisterResponse;
 import com.playkuround.playkuroundserver.domain.user.exception.UserEmailDuplicationException;
 import com.playkuround.playkuroundserver.domain.user.exception.UserNicknameDuplicationException;
+import com.playkuround.playkuroundserver.domain.user.exception.UserNicknameUnavailableException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.ValueSource;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
@@ -20,7 +23,7 @@ import org.springframework.test.util.ReflectionTestUtils;
 import java.util.Date;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.when;
 
@@ -86,21 +89,35 @@ class UserRegisterServiceTest {
 
             // expect
             UserRegisterDto userRegisterDto = new UserRegisterDto(email, nickname, Major.valueOf(major));
-            assertThrows(UserEmailDuplicationException.class,
-                    () -> userRegisterService.registerUser(userRegisterDto));
+            assertThatThrownBy(() -> userRegisterService.registerUser(userRegisterDto))
+                    .isInstanceOf(UserEmailDuplicationException.class);
         }
 
         @Test
         @DisplayName("이미 존재하는 닉네임이면 에러가 발생한다.")
-        void signupFailByExistsNickname() {
+        void fail_2() {
             // given
             when(userRepository.existsByEmail(email)).thenReturn(false);
             when(userRepository.existsByNickname(nickname)).thenReturn(true);
 
             // expect
             UserRegisterDto userRegisterDto = new UserRegisterDto(email, nickname, Major.valueOf(major));
-            assertThrows(UserNicknameDuplicationException.class,
-                    () -> userRegisterService.registerUser(userRegisterDto));
+            assertThatThrownBy(() -> userRegisterService.registerUser(userRegisterDto))
+                    .isInstanceOf(UserNicknameDuplicationException.class);
+        }
+
+        @ParameterizedTest
+        @ValueSource(strings = {"존나", "개새끼", "씨발", "시발"})
+        @DisplayName("BadWordCheck에 걸리는 닉네임은 에러가 발생한다.")
+        void fail_3(String badWord) {
+            // given
+            when(userRepository.existsByEmail(email)).thenReturn(false);
+            when(userRepository.existsByNickname(badWord)).thenReturn(false);
+
+            // expect
+            UserRegisterDto userRegisterDto = new UserRegisterDto(email, badWord, Major.valueOf(major));
+            assertThatThrownBy(() -> userRegisterService.registerUser(userRegisterDto))
+                    .isInstanceOf(UserNicknameUnavailableException.class);
         }
     }
